refactored header and converted components to tsx
