+++ app/components/DonateTrees/screens/DonorDetails.native.js
@@ -18,10 +18,9 @@ import { updateStaticRoute } from "../../../helpers/routerHelper";
 import i18n from "../../../locales/i18n.js";
 import styles from "../../../styles/donations/donorDetails";
 import { formatNumber } from "../../../utils/utils";
-import {
-  SelectCountryModal,
-  Header
-} from "../components/donationComponents.native";
+import { SelectCountryModal } from "../components/donationComponents.native";
+import { Header } from "./../components/Header";
+
 import countryData from "../../../assets/countryCodes.json";
 import SafeAreaView from "react-native-safe-area-view";
