+++ package.json
@@ -161,6 +161,8 @@
     "@babel/preset-env": "^7.8.6",
     "@babel/preset-react": "^7.8.3",
     "@babel/runtime": "^7.8.4",
+    "@types/react": "^16.9.41",
+    "@types/react-native": "^0.62.14",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
@@ -187,12 +189,14 @@
     "prettier": "1.19.1",
     "prettier-eslint": "^9.0.1",
     "react-native-global-props": "^1.1.5",
+    "react-native-typescript-transformer": "^1.2.13",
     "react-test-renderer": "16.12.0",
     "reactotron-react-js": "^3.3.7",
     "reactotron-react-native": "^4.0.3",
     "reactotron-redux": "^3.1.2",
     "sass-loader": "^8.0.2",
     "terser-webpack-plugin": "^2.3.5",
+    "typescript": "^3.9.5",
     "url-loader": "^3.0.0",
     "webpack": "^4.41.6",
     "webpack-bugsnag-plugins": "^1.4.3",
